Corretti bug. Da testare. Issue 9:	Invio comandi senza accorpamento per nodo e senza rispetto frequenza di invio Issue 8:	T22 - Nuovo stato a NULL in aggiornamento

+++ src/main/java/org/openhab/binding/souliss/internal/network/typicals/SoulissT11.java
@@ -72,7 +72,7 @@ public class SoulissT11 extends SoulissGenericTypical {
 	
 	@Override
 	public State getOHState() {
-		String sOHState=StateTraslator.statesSoulissToOH(this.getNote(), this.getType(),this.getState());
+		String sOHState=StateTraslator.statesSoulissToOH(this.getNote(), this.getType(),(short) this.getState());
 		if(sOHState!=null){
 		if (this.getNote().equals("ContactItem"))
 			return OpenClosedType.valueOf(sOHState);
